Small refinement to the auto-decompressing-execs. Avoid a zombie.
The script forked a process to delete the tmp file and then execed the decrompressed bynary. That left the "sleep-then-rm" process behind as a zombie.

Instead do the opposite: the script forks a process to exec the decrompessed binary and stays behind to unlink the file. (Also gets the status of the exec child so it can report an error if it died early). The child, if still alive, gets adopted by init, like any other system daemon.