A better improvement to the self decompressing exec script.
The previous attempt was cheating too much. In worked nicely for daemons but not at all for interactive commands.

This one goes back to running the cleanup in a child. It prevents the zombification of that child by making it a grand-child. The child terminates immediately, resulting in the grand child being orphaned and so adopted by init.