doc: fixup certificate copying in deploy tutorial
Fix copying of certificates to individual hosts in the "freestanding
deployment" tutorial.
Was suggesting to copy all certificates to `/etc/scion/crypto/`.
Instead, it should copy the AS certificates to `/etc/scion/crypto/as/`.

Originally the instructions had been correct; the problem was introduced
with the custom script to generate the crypto material. This script uses
flatter paths (for brevity) than the originally used `scion-pki testcrypto`
tool, and the copy instruction were not correctly updated (and not tested, duh).