It is ok to remove versioning.bzl after all.
Having a version in the repo doesn't do much good. It is a rare thing that the
build fails for lack of this file as any bazel execution (that doesn't need it
i.e. not a dist-* target) will generate it. Better get the error in the rare
case of building a dist- target as the very first bazel command than using the
wrong version number or the constantly dirty version due to the file being
tracked by git.