packaging: Take the version tag for file names from the command line.
At the request of the reviewers, no-longer auto-generate a version tag for file
names. This is to avoid the need for a bazel wrapper script or other acrobatics.
It happens that we also rarely need the versionned names; normally only when
building publishables from the CI pipeline; so an extra argument on the
command line isn't inconvenient.

We now rely on the pkg_deb support for configurable file names and have added
similar support in openwrt packaging rules. versioning.bzl, tools/bazel, and
the file copies in the Makefile have been replaced by a simple symlink of all
assets to a convenient location.

In the absence of directives, the version tag is "dev".
The version strings embeded in packages and binaries is always generated, as
before.