Fix performance issue.
Fix may be an exageration here. The root cause is not known but correlates with
using the assigned mac address of the brload-side interface. This only affects
the regression testing context; local processes communicating via a veth
pair. Rough hypothesis: incoming packets that have the "correct" destination
MAC address get processed by the kernel, others do not. The processing is
entirely unnecessary as brload uses a raw socket for everything. The
processing is expensive (net filter?) and is also done while blocking the
sender.

Ways of bypassing this processing (besides the address trick) are still being
researched.