testing: added performance index to benchmark report (of external router).
This involves incorporating the mbw and coremark microbenchmarks into the
bmtools package, run them at start/install time, collect the output into
a json file that be scraped by the benchmark driver.

The benchmark driver now scrapes that and uses it to compute the perf index.

Caveat: it appears that mbw, if compiled for musl libc (as opposed to
statically against glibc), performs very poorly. So musl_libc's memcpy
is about 1/2 the speed of libc's. Looking for a fix. Most likely, write
an equivalent in Go, so, at least we test the memcpy that we most use.