Replaced the mbw benchmark with a Go equivalent.
The problem with mbw is the C library. That benchmark looks at performance from
the application's p-o-v, so measures a vanilla memcpy. There are two issues
with that:
* our router is written in Go, using its own copy function. So memcpy isn't
  all that relevant.
* the memcpy implementation of musl_libc is terrible (1/2 speed). This means
  that we measure a memory bandwith badness that has no effect whatsoever on
  our router.

By using a go program, we get consistent results. i.e. differences in result
reflect differences in the hardware platform, not differences in the c library.