cmd/scion, snet: reduce range of IDs for SCMP requests (#4495)
Preparation for dispatcher-less endpoints. Reduce the range of IDs used
for SCMP requests, to be forward-compatible with the default range for
port-dispatching in the router.

In the dispatcher-less future, the router will forward SCMP reply
messages to an application's underlay port determined based on the SCMP
message ID. For this, the router will have a configurable range of ports
for which it applies the dispatching, allowing dispatcher-ful and
dispatcher-less endpoints to coexist. This port range will need to be
disjoint from the port range currently used by the dispatcher and SCMP
request identifiers.

For more details, see design document, #4280.