cppki/segverifer: enforce AS certificate to cover AS entry (#4492)
When verifying an AS entry the verifier will now only use an AS
certificate that is valid until the expiration of the AS entry. Before,
the verifier would use any certificate that was valid at the time of
verification but not necessarily for the entire validity of the AS
entry, i.e., the hop field expiration contained in the entry.
The verifier now requests certificates that cover at least the entire
validity period of the AS entry. To that end, the `ChainsRequest.date`
field has been renamed to `ChainsRequest.at_least_valid_until` and a new
`ChainsRequest.at_least_valid_since` field has been introduced.
Legacy clients of the API will only set the `at_least_valid_until` field
(for them it's still the `date` field, but the name of the field does
not matter for protobuf wire enconding). New servers will simply set
`at_least_valid_since` to be equal to `at_least_valid_until` to keep the
same semantics.
Legacy servers will ignore the `at_least_valid_since` field, since they
don't know about it.
This only enforces that the AS certificate is still valid at the AS
entry expiration, but not that it was also valid at the start of the hop
field validity. Once all trust material servers and clients have been
upgraded we automatically have the desired behavior.

Fixes #4286