private/env: remove RunsInDocker during startup log (#4499)
Remove the call to RunsInDocker on application start up and remove the
corresponding log line, "In docker: <bool>".
As the implementation of this check is Linux-specific, it would always
error on non-Linux platforms and exit the program. Because the
usefulness of this log line is questionable, it was removed entirely
instead of fixing the problem by e.g. only logging this when running on
Linux.