Updated the benchmark based on observations.
There were a number of ways in which the benchmark failed to reflect reality
on the platforms we know:

* Working set too large, intended to defeat cache: not realistic. For example
  our router buffers, in the benchmark conditions, fit in the cache. Made
  the working set comparable in size to that of the router.
* Incorrect copy direction. We were copying uninitialized data onto initialized
  one instead of the opposite. Fixed.
* Copy order, we were copying all buffers sequentially in the forward direction
  which rewards prefetching to the unrealistic maximum extend. Replaced with
  random copies between buffers.

This yields very high numbers for modern platforms, like my laptop or the CI
system, and very low for an old embeded system like the APU2. This is much more
in line with the observed raw packet rates. To such an extent that they could be
predicted solely based on that number.