testing: extend router benchmark tooling (#4496)
Enables the benchmarking of a SCION router running on an external
device.

What this change does is to simplify the task of executing the same performance benchmark as done by the CI but on
a physical router. It provides a script that:

* Instructs the user to configure (interfaces, SCION router config and topology) and wire the subject router appropriately for the benchmark.
* Configures the local host's interfaces appropriately for simulating the subject's neighboring routers.
* Executes the benchmark.
* Collects some hardware performance characteristics from the subject or from the command line, if available.
* Outputs a report.

This change also includes an openwrt package which, when installed sets the subject router's SCION configuration automatically. Leaving the user with only the task of configuring and cabling two interfaces. For other platforms the configuration is still manual.

---------

Co-authored-by: Matthias Frei <matzf@scion.org>