Fix dataplane_test after merge with dispatcher removal.
Also, found a bug that was revealed by the dispatcher removal changes in the router:
One of the tests sends a raw payload with no SCION/UDP header but was crafting messages that
indicated a SCION/UDP header. Since the dispatcher removal PR, the router snoops in L4 and
mistakes the payload for a UDP header. That is invisible unless the payload is empty.\
No idea why the test doesn't break in master but it brakes in this PR.