pkg/metrics: introduce v2
Introduce a v2 version of pkg/metrics. The current metrics package is
not optimal:
- Using objects that have been called with With allocated on
  each increment/set value call.
- Having the option to call With ties consumer code with initialization
  code.
- There is no way whether a counter is fully initialized with all
  labels.

A full discussion can be found in
https://github.com/scionproto/scion/pull/4527.

The new v2 version is compatible with prometheus, so no explicit
conversion functions are offered, instead the prometheus creation
functions can be used directly.

So far the new library has been plugged in selected places. The
intention is that it will eventually replace the current library
completely, however doing so in a single sweep would be too extensive.