doc: fixup generated documentation files after removing dispatcher flags (#4529)
Run `bazel run //:write_all_source_files` to fix up generated documentation after removing the dispatcher flag from scion/scion-pki binaries.

This was missed during the PR because the existing check for this was not active in the CI. Extend the bazel configuration `test:lint` to enable running this in `make lint`.