connectrpc: generate code with buf
Generate the additional code required for connectrpc with the buf
tool. Currently, there is no bazel ruleset available for buf. We work
around this by invoking the buf binary directly and commiting the
generated code in git. To deal with generation, we pin the
`protoc-gen-connect-go` version in the go.mod file. A small scaffolding
go binary is added to deal with setting up the environment correctly
for the buf tool. While this could be a shell script, this is more
readable and maintainable.

This PR is broken out of the
https://github.com/scionproto/scion/tree/connectrpc branch to make
incremental review easier.

TL;DR: Run `bazel run //tools/buf` to generate the connectrpc code.