Cleanup the dispostion/error duality in processPacket.
In ProcessPacket the error value is used very little and we sometimes build it
at great expense. Replaced all of that with just returning the packet
disposition (Discard, Forward, or Slowpath) everywhere. The error, when there
is one is reduced to one of the preallocated value, with no context, and is
logged on-the-spot, instead of being passed up the stack just for the purpose
of logging it.