build: update rules_js and pnpm version
Updated tooling allows to use lockfile version 9 in pnpm-lock.yaml.
We've recently seen that dependabot PRs for dependency updates seem to
create lock files with that lockfile version; as the tooling did not
support it, the dependabot changes could not be used. Plausibly,
dependabot PRs will work after this change.

Note that pnpm cannot be upgraded to the very latest version 9.5.0, as
this needs to be explicitly enabled in aspect_rules_js first.

Note that I've explicitly "downgraded" the nodejs node_version to the
default version used in the previous rules_nodejs version. The default
version of the current rules_nodjs version requires a too new glibc for
my ancient setup.