Cleanup use of snet.UDPAddr from ping and traceroute. (#4573)
This type was being abused to represent sources and destinations
addresses that are not UDP and do not have a port. Now we have addr.Addr
to represent SCION addresses independently of UDP. Ping and traceroute
were taking advantage of the type's fields to convey paths and next hop
addresses internally. Those values are now passed as parameters or
stored in internal data structures.

Fixes #4275