serrors: delete error context merging and code cleanup (#4586)
The withCtx() function, in rare cases, created hybrid objects that required complicated support code and, in
some cases could cause Is() to panic.

Removed the feature, replaced WithCtx() with a shim that doesn't create
hybrids. Removed all the complicated code that resulted.
There are now two subclasses of errors:
* basicError: dedicated to errors of the form string+cause - Constructed
by Wrap()
* joinedError: dedicated to errors of the form error+cause - Constructed
by Join()
This allows for Unwrap() to be different and do the right thing, so that there is no need for a custom implementatin of Is().

Introduced new constructors with obvious semantics to replace those already marked deprecated. Also replaced the ctx map by a slice of ctx pairs. Which removes some unnecessary conversion code.

I still plan on editing the calls to deprecated constructors (and common.ErrMsg) out of everywhere, but in a subsequent PR.

Fixes #4486