brload: fix premature socket closure (#4591)
Brload borrows the file descriptor from a af_packet.TPacket structure in
order to use it with sendmmsg (not available via TPacket). In the
process it left the original TPacket, only used on the receiving side,
exposed to garbage collection, resulting in the premature closure of the
socket. Fixed it by keeping a reference to TPacket on the sending side.

In passing:
* increase the benchmark packet size to 1500 bytes, to show
bandwidth (~10Gb/s) with normal frames.
* display brload's output to help in diagnosing issues.

Fixes #4590