Bump go version to 1.23
This required upgrading rules_go to 0.47.0 because of some magic
parameter that must now be passed to the go assembler in support of the
internal/V0 ABI split in go 1.22.

It wasn't possible to upgrade rules_go further without also upgrading bazel and
other stuff. So... one problem at a time.

As a result of rules_go upgrade, the proto generation has changed, hence the
other diffs.