docs: add interactive TRC signing ceremony builder
And an interactive TRC signing ceremony builder.
It supports all three types of TRC ceremony: base, regular, and sensitive.
The user can select between scion-pki and openssl.
Furthermore, support for pkcs11 is available for openssl.
In a future iteration, support for scion-pki kms will be added.

The builder is intended to be used by a voting (or root CA) participant.
The adminstrator role is not supported in this version. (Administrators
should be very knowledgable about the TRC ceremony anyway.)