test: await_connectivity could complete prematurely.
It seems (by reading the code and trying things by hand) that await_connectivity could be satisfied by core segments while looking for up-segments.

So, I made the part checking for leaf connectivity more picky. It also waits for the matching down segment to be registered (although, contrary to what I thought, it happens quickly).

That seems to negate the need for an extra 10s delay, so removed it.