tutorial: topology source host missing causes scion-router@br.service to crash (#4619)
Hi,

I've tried to setup the freestanding SCION network as per the
[tutorial](https://docs.scion.org/en/latest/tutorials/deploy.html) and
ran into a few issues:

Using the default topology files from the website, the scion-router
service always crashes because of an invalid source IP (see screenshot
below):


![image](https://github.com/user-attachments/assets/cef3fa72-a945-438f-ac31-1bc535da2a44)

@tobiasmoldan realized that this can be resolved by adding the source IP
(or host name) to the local field in the topology:

`"underlay": {
            "local": "scion01:50012",
            "remote": "scion02:50012"
          },`

Secondly, I was wondering if all ASes need to be up and running for the
network to be fully functional? I have set up and configured scion01 and
scion02 as per the tutorial. All services (daemon, router, control,
dispatcher) are active and there is a direct link between the two VMs.
The machines are pingable (via normal ICMP), however, there is no
reachability via SCMP. I assume that, even without scion03-05, they
should have connectivity?

I got scion addresses for scion01 and scion02 respectively,

`42-ffaa:1:1,127.0.0.1`

`42-ffaa:1:2,127.0.0.1`

and I have generated the certificate material on scion01 and shared it
with scion02. Is there anything I'm missing - e.g., does path
exploration need to be triggered somehow or should it happen
automatically?

Thanks a lot for your help on this!

Best regards,

Thomas

---------

Co-authored-by: jiceatscion <139873336+jiceatscion@users.noreply.github.com>