test: await-connectivity is too optimistic (#4622)
Found this in the wake of #4606
I believe that await-connectivity could mistake core segments for up
segments (i.e. assuming that only up segments could be found). It still
makes the optimistic assumption that down segments are registered
immediately after up segments are obtained. We have to be content with
that because in hidden paths test cases the down segments cannot all be
found via a simple REST API query.