Fix python lint.
For some very mysterious reasons, flake8 gets broken by build_python_zip.
I have never been able to find how exactly flake8 is made as a py binary
because its creations is mediated by 6 layers of bazel code generation. But
it is indeed made during the build.

If I turn build_python_zip ON for the "build" config only, then flake8 is still
broken, but if I turn it ON everywhere except for the "test:lint" config,
then flake8 works. That makes no sense to me, but I'll be content with that.
Life is short and I am sick of spending mine trying to deconstruct bazel
and python frivolous and broken magic.