build: do not run pip install outside the bazel build framework (#4647)
I was sick and tired of stepping on that Lego every time I built on a
new machine...

make mocks, was running mocks.py directly instead of calling bazel run.
That used whatever ambient python and libraries to match instead of the
python env specified by WORKSPACE, which meant that we had to include
pip install in the tools/install_deps process, which often enough would
fail or require no-end of messy work-arounds (e.g. create a python venv
and remember to activate it before building - but just the mocks).