router: small refactoring to remove the need for a batchconn.WriteTo() API.
WriteTo needs to be given the destination address explicitly. WriteBatch, on
the other end, can either find it in each packet structure, or rely on the
connection's destination. WriteTo is only used to send BFD packets. It turns
out that BFD packets can also very easily be sent via the regular forwarders
that use WriteBtach.

The motivation to do that is to simplify the interface between the dataplane
and the forwarders, in view of supporting multiple underlays with possibly
very different interfaces. So the channels around the processor tasks seems
like a good universal interface.