build: work around issues between pkg_rpm and rpmbuild after 4.20 (#4652)
Since rpmbuild v4.20, the _builddir variable is forcefully set to
<chosen_builddir>/<pkg_name> (which is created if it doesn't exist). Any
files that are not built/copied by the %prep% phase must be already
present in <chosen_builddir>/<pkg_name>/[%buildsubdir].

pkg_rpm doesn't expect that and places the files to be packaged in
<chosen_builddir>. Since %buildsubdir is not used (=="."), we compensate
by setting it to "..". As a result
<chosen_builddir>/<pkg_name>/[%buildsubdir] evaluates to
<chosen_buildir> thereby restoring pre-4.20 behavior.

Since 4.20, debug-info rpms are built by default. We don't support or
want that, so we now set "don't-do-it" explicitly.

Fixes #4653