pkg/snet: ignore decoding/parsing reading errors on SCIONPacketConn (#4670)
So far, the `snet/SCIONPacketConn` was bubbling up decoding/parsing
errors to the calling applications. In principle, these errors are not
recoverable by the application. The current solution simulates that the
network stack simply discards incorrect/malformated packets.
Otherwise,some libraries/applications would misbehave (similar to what
would happen if were to receive SCMP errors). If we believe that
SCION-aware applications using the low-level `SCIONPacketConn` should
receive the error, we can move this up to `Conn`.