Move a large part of underlay-specific code out of the main router.
The underlay code is now the udpip underlay provider.

The refactoring is not complete. The assumption that underlay addresses are
udp/ip addresses is still part of the router. Also, the main receiver and
sender tasks are still where they are and use the underlay connections via
a temporary API. The reason for that is to make the diff somewhat more
reviewable by not moving large swaths of code from one file to another.