drkey: simplify Epoch type (#4690)
During verification of the drkey package (as part of the `VerifiedSCION`
project) we ran into an issue with our verifier. It complains about the
embedded struct in `Epoch`, which in this specific case, does not seem
to be supported.

(Just not to confuse anyone: The issue is with the verifier (gobra), not
with Go or the `drkey` package itself.)

However, to me there does not seem to be a reason why `Epoch` embedds
`cppki.Validity`. It could as well just be a type definition. This is
what I propose with this PR. It changes the definition of `Epoch` and
adapts (and simplifies) the clients of the `drkey` package.

---------

Co-authored-by: aaronbojarski <abojarski@student.ethz.ch>