router: make room for multiple underlay impls phase 1 (#4658)
The goal is to eventually remove all underlay-specific code from the
main router code and place it in a plugin, such that there can then be
multiple underlay plugins. This PR isn't attempting to fully accomplish
that. It avoids moving some code, so that the diff is easier to read.
The price to pay is a few extra spaghetti left between main code and
plugin.

The code that should eventually move to the underlay is:
* runReceiver and runForwarder.
* likely some of the BFD code.
* opening of connections (currently done by connector.go)

Other changes being planned:
* Stop reusing the internal connection for sibling links (so we can take
advantage of bound connections).
* Add knowledge of multiple underlay into the configurator.
* Make underlay addresses opaque to the router.
* Demultiplex to links on ingest, so ifID and srcAddress (when they are
defined by the link) are obtained in the most efficient way (directly
from the link's fields for example).

Contributes to: #4593