router: add tcp support for dispatching to endhost (#4703)
This PR resolves #4698. 

I followed the same principle that we use for the other protocols, i.e.,
we dispatch to the mapped port if the destination port is within the
dispatched port range. Otherwise, it is sent to the fix `EndhostPort`.

This makes it compatible with implementations following the
"Wireguard"-model, i.e., sending the encapsulated traffic over a fix UDP
port, e.g., the [SCION-IP translator](
https://github.com/netsys-lab/scion-ip-translator); or alternative
implementations that map the underlay port to the application port,
e.g., what we do with SCMP or UDP/SCION, provided proper
`dispatched_ports` configuration.