Push Host address resolution down into the underlay.
Since resolution is done only on the internal network, and since the internal
network is still assumed to be udpip, address resolution was moved to the
udpip underlay and is invoked with interfaces[0].ResolveHostPort.

The rest is pretty much unchanged. I had to share two sentinel errors b/w
router and underlay. I had to move the dispatched port range configuration to
the underlay as well. Fair, since it is quite uniquely a udp/ip feature and
also strictly meant for the internal network.

The last knowledge of underlay address outside of the underlay is for service
resolution. That's next.