Associate interface metrics with links, not ifIDs. Also update some comments.
The reason for the interface metrics change is that the sibling links are
distinct, and statistics about them are useful. However, since they do not
have an ifID, it is not possible to map their statistics by ifID. It is also
not particuly expedient as it requires an extra [id]metric map. We still cannot
sort transit packet metrics by remote external interface, but that's still
an improvement. Finaly the difficulty in dealing with the map by ifid was
creating some inconsistencies.