router: refactor phase 4 (#4737)
Made the router almost entirely oblivious of underlay addresses. The
router now understands only
one underlay address: its own local address on the internal network (and
still relies on the assumption that the internal link is on the udpip
underlay.)

The router still provides storage for the remote address in the packet
structure, which is a net.UdpAddr. It does not use it but resets it when
taking a free packet from the pool. This will take more effort to
eliminate.

Contributes to: [#4593](https://github.com/scionproto/scion/issues/4593)